+++ @navikt/core/react/src/data/token-filter/AutoSuggest.types.ts
@@ -1,0 +1,14 @@
+interface OptionGroup<T> {
+  label: string;
+  options: T[];
+}
+
+interface AutoCompleteOption {
+  value: string;
+  label: string;
+  tags?: string[];
+  filteringTags?: string[];
+  description?: string;
+}
+
+export type { AutoCompleteOption, OptionGroup };
